Transforming BASEX to a standalone package added setup.py
