+++ setup.py
@@ -1,0 +1,14 @@
+#!/usr/bin/python
+
+from setuptools import setup
+
+
+setup(name='BASEX',
+      version='0.1',
+      description='A Python implementation of the BASEX algorithm',
+      author='Dan Hickstein',
+      packages=['BASEX'],
+      package_data={'BASEX': ['BASEX/data/*','BASEX/tests/data/*' ]},
+      #test_suite="BASEX.tests.run"
+     )
+
